requirements.txt
Flask==3.0.0
pandas==2.2.2
openpyxl==3.1.2
xlrd==2.0.2
pyxlsb==1.0.10
reportlab==4.0.7
Werkzeug==3.0.1
numpy==1.26.4
Flask-Session==0.5.0
supabase==2.3.0
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
lxml==5.3.0
